refactor: product dto 분리(request, response)  (#5)
