+++ src/main/java/com/kt/dto/product/request/ProductUpdateCategoryRequest.java
@@ -1,4 +1,4 @@
-package com.kt.dto.product;
+package com.kt.dto.product.request;
 
 public record ProductUpdateCategoryRequest(
 	Long categoryId
