+++ src/main/java/com/kt/dto/product/request/ProductUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.kt.dto.product;
+package com.kt.dto.product.request;
 
 import jakarta.validation.constraints.Min;
